Define Method scan()
This method should read one character at a time and return the TOKEN enum value
corresponding to the token that was just read. It should add characters to
the token buffer as necessary. When scanning an ID (an ID is a variable
name) the lexeme for the ID should be in the token buffer and it should
return ID as the token value at the end of the method. All whitespace
appearing in the input should be ignored.